[SPARK-43428] Move some class utils to common/utils
