+++ common/utils/src/main/scala/org/apache/spark/SparkThrowableHelper.scala
@@ -22,6 +22,7 @@ import scala.collection.JavaConverters._
 import com.fasterxml.jackson.core.util.MinimalPrettyPrinter
 
 import org.apache.spark.util.JsonUtils.toJsonString
+import org.apache.spark.util.SparkClassUtils
 
 private[spark] object ErrorMessageFormat extends Enumeration {
   val PRETTY, MINIMAL, STANDARD = Value
@@ -33,7 +34,7 @@ private[spark] object ErrorMessageFormat extends Enumeration {
  */
 private[spark] object SparkThrowableHelper {
   val errorReader = new ErrorClassesJsonReader(
-    Seq(getClass.getClassLoader.getResource("error/error-classes.json")))
+    Seq(SparkClassUtils.getSparkClassLoader.getResource("error/error-classes.json")))
 
   def getMessage(
       errorClass: String,
